html/msm_marketing_ai/requirements.txt
praw==7.7.1
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
Pillow==10.0.0
moviepy==1.0.3
instabot==0.117.0
TikTokApi==5.2.2 